discord_bot_mcp/pyproject.toml
[project]
name = "discord-mcp-server"
version = "0.1.0"
description = "Model Context Protocol server for Discord integration"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.32.1",
    "fastapi>=0.100.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["discord_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=discord_mcp --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
discord-mcp-server = "discord_mcp.server:main"
